backend/requirements.txt
fastapi
uvicorn
python-dotenv
opencv-python
ultralytics
websockets
